Update f194c2fe-9c00-436e-a3ab-250764a9aae6.md
I added return types and fixed the obvious typos. I am testing this example right now and really hope it would world later. As far as I can see the TwigDependencyProvider is part of this service and provides the method to register. I have to no how this is discovered by the system. And  a full implementation of this class would be an amazing addidtion as this is not an extension of Spryker core as the other plugin providers.