Update bdb56333-569c-42ac-9a12-2f8ecc84c6b5.md
Changed text of line 138 from: "Calculation defines the value to be deducted from a product's original price. There are two types of discount calculation:"
To: "Calculation defines the value to be deducted from a product's default price. There are two types of discount calculation:"

Discount calculation uses the product's default price, not its original price. I confirmed this with another PM and a developer familiar with discount calculation.